Add common header
